\title{IsoGeneGUI Help}
\name{IsoGeneGUIHelp}
\alias{IsoGeneGUIHelp}
\description{
Function to for opening the IsoGeneGUI help.
}

\usage{
IsoGeneGUIHelp()
}

\details{
To run the package, we use the function: IsoGeneGUI().
}

\author{
Author: Setia Pramana, Dan Lin, Philippe Haldermans, Tobias Verbeke, Martin Otava
Maintainer: Setia Pramana <setia.pramana@ki.be>
}


\keyword{documentation}

\examples{
\dontrun{
library(IsoGeneGUI)
IsoGeneGUIHelp()
}
}
