\title{IsoGeneGUI}
\name{IsoGeneGUI}
\alias{IsoGeneGUI}

\description{
This function will load the IsoGeneGUI package.
}


\usage{
IsoGeneGUI()
}


\details{
To run the package, we use the function: IsoGeneGUI().
}
\author{
Author: Setia Pramana, Dan Lin, Philippe Haldermans, Tobias Verbeke, Martin Otava

Maintainer: Setia Pramana <setia.pramana@ki.be>
}


\keyword{documentation}
\examples{
\dontrun{
library(IsoGeneGUI)
IsoGeneGUI()
}
}
